cooldown for bullets bug fix

+++ Shadow Adventure/src/gravity.cpp
@@ -17,11 +17,15 @@ void gravitationB(BMAP* bullets, std::vector<std::vector<char>>& board) // gravi
 		{
 			if (entity.getGravity()) // if bullet have gravity
 			{
-				board = entity.kill(board); // kill bullet
-				entity.move(entity.getGType()); // move bullet
-				board = entity.spawn(board); // spawn bullet
-				nbm.insert(std::make_pair(entityS.first, entity)); // add bullet to new bullets map
+				if (!(entity.getC() > 0))
+				{
+					board = entity.kill(board); // kill bullet
+					entity.move(entity.getGType()); // move bullet
+					board = entity.spawn(board); // spawn bullet
+					entity.setC();
+				}
 			}
+			nbm.insert(std::make_pair(entityS.first, entity)); // add bullet to new bullets map
 		}
 	}
 	bullets->set(nbm); // set old bullets map
